# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFtaXJleDEyOEBnbWFpbC5jb20iLCJleHAiOjUyNjgzMTg5OTYsImV4cGlyZV9hdCI6IiIsImZpcnN0bmFtZSI6Itin2YXbjNixIiwiaWQiOjEsImxhc3RuYW1lIjoi2LTbjNix2K_ZhNuMIiwibW9iaWxlIjoiMDkwMjQ4MDk3NTAiLCJvcmlnX2lhdCI6MTY2ODMyMjU5Niwic3RhdHVzIjoiIn0.x7BKuxw288cm1JsskGRD178UPmNz-xRwkWHtb0WsU74",
        "title": "Selloora Backend API",
        "contact": {
            "name": "API Support",
            "url": "https://www.amirshirdel.ir",
            "email": "amirex128@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8585",
    "basePath": "/",
    "paths": {
...
        "/user/gallery/create": {
            "post": {
                "description": "با آپلود یک تصویر میتوانید شناسه آن را در بخش های مختلف استفاده نمایید و در آینده بر اساس همین شناسه تصویر را حذف نمایید همچنین تمامی تصاویر به فرمت وب پی تبدیل میشوند",
                "tags": [
                    "gallery"
                ],
                "summary": "آپلود تصویر",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.CreateGallery"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/gallery/delete/{id}": {
            "post": {
                "description": "حذف یک تصویر از گالری",
                "tags": [
                    "gallery"
                ],
                "summary": "حذف گالری",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "شناسه گالری",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/menu": {
            "get": {
                "description": "با ایجاد منو کاربر میتواند منو های بالای صفحه و پاین صفحه مربوط به قالب خود را کم و زیاد نماید",
                "tags": [
                    "menu"
                ],
                "summary": "لیست منو ها",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "متن جستجو",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "شماره صفحه",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "تعداد صفحه",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "مرتب سازی براساس desc/asc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/user/menu/create": {
            "post": {
                "description": "با ایجاد منو کاربر میتواند منو های بالای صفحه و پاین صفحه مربوط به قالب خود را کم و زیاد نماید",
                "tags": [
                    "menu"
                ],
                "summary": "ایجاد منو",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.CreateMenu"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/menu/delete/{id}": {
            "post": {
                "description": "با ایجاد منو کاربر میتواند منو های بالای صفحه و پاین صفحه مربوط به قالب خود را کم و زیاد نماید",
                "tags": [
                    "menu"
                ],
                "summary": "حذف منو",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "شناسه منو",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/menu/show/{id}": {
            "get": {
                "description": "با ایجاد منو کاربر میتواند منو های بالای صفحه و پاین صفحه مربوط به قالب خود را کم و زیاد نماید",
                "tags": [
                    "post"
                ],
                "summary": "نمایش منو",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "شناسه",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/menu/update": {
            "post": {
                "description": "با ایجاد منو کاربر میتواند منو های بالای صفحه و پاین صفحه مربوط به قالب خود را کم و زیاد نماید",
                "tags": [
                    "menu"
                ],
                "summary": "ویرایش منو",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.UpdateMenu"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "DTOs.CreateGallery": {
            "type": "object",
            "required": [
                "height",
                "width"
            ],
            "properties": {
                "height": {
                    "description": "طول تصویر",
                    "type": "integer",
                    "example": 300
                },
                "width": {
                    "description": "عرض تصویر",
                    "type": "integer",
                    "example": 300
                }
            }
        },
        "DTOs.CreateMenu": {
            "type": "object",
            "properties": {
                "link": {
                    "description": "لینک منو",
                    "type": "string",
                    "example": "https://example.selloora.com/page/test"
                },
                "name": {
                    "description": "نام منو برای نمایش در لیست ها",
                    "type": "string",
                    "example": "خانه"
                },
                "parent_id": {
                    "description": "شناسه منو بالا سری برای منو های اصلی و باید صفر ارسال شود",
                    "type": "integer",
                    "example": 0
                },
                "position": {
                    "description": "محل قرار گیری منو در قالب که با توجه به تنظیمات قالب میباشد",
                    "type": "string",
                    "enum": [
                        "top",
                        "bottom",
                        "right",
                        "left"
                    ],
                    "example": "top"
                },
                "shop_id": {
                    "description": "شناسه فروشگاه",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "DTOs.UpdateMenu": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "شناسه منو برای ویرایش",
                    "type": "integer",
                    "example": 1
                },
                "link": {
                    "description": "لینک منو",
                    "type": "string",
                    "example": "https://example.selloora.com/page/test"
                },
                "name": {
                    "description": "نام منو",
                    "type": "string",
                    "example": "خانه"
                },
                "parent_id": {
                    "description": "شناسه منو بالا سری",
                    "type": "integer",
                    "example": 0
                },
                "position": {
                    "description": "شناسه منو بالا سری برای منو های اصلی و باید صفر ارسال شود",
                    "type": "string",
                    "enum": [
                        "top",
                        "bottom",
                        "right",
                        "left"
                    ],
                    "example": "top"
                },
                "sort": {
                    "description": "ترتیب منو",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}